+++ src/capitalize_decorator.rb
