+++ src/decorator.rb
