+++ src/nameable.rb
