+++ src/person.rb
@@ -1,9 +1,5 @@
 require_relative 'nameable'
 
-require_relative 'capitalize_decorator'
-
-require_relative 'trimmer_decorator'
-
 class Person < Nameable
   attr_reader :id
   attr_accessor :name, :age
@@ -28,10 +24,3 @@ class Person < Nameable
     @name
   end
 end
-
-person = Person.new(22, 'maximilianus')
-puts person.correct_name
-capitalized_person = CapitalizeDecorator.new(person)
-puts capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-puts capitalized_trimmed_person.correct_name
