--- trimmer_decorator.rb
@@ -1,7 +1,0 @@
-require_relative 'decorator'
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable[0..9]
-  end
-end
